+++ apps/fabric-website-resources/src/components/pages/FacepilePage.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { DemoPage } from '../DemoPage';
 
-import { FacepilePageProps } from 'office-ui-fabric-react/lib/components/Facepile/Facepile.doc';
+import { FacepilePageProps } from '@fluentui/react-examples/lib/office-ui-fabric-react/Facepile/Facepile.doc';
 
 export const FacepilePage = (props: { isHeaderVisible: boolean }) => (
   <DemoPage