+++ packages/fluentui/react-northstar/src/themes/teams-high-contrast/componentVariables.ts
@@ -41,3 +41,4 @@ export { dropdownSelectedItemVariables as DropdownSelectedItem } from './compone
 export { labelVariables as Label } from './components/Label/labelVariables';
 export { tooltipContentVariables as TooltipContent } from './components/Tooltip/tooltipContentVariables';
 export { datepickerCalendarCellVariables as DatepickerCalendarCell } from './components/Datepicker/datepickerCalendarCellVariables';
+export { datepickerCalendarCellButtonVariables as DatepickerCalendarCellButton } from './components/Datepicker/datepickerCalendarCellButtonVariables';