+++ packages/office-ui-fabric-react/src/components/Checkbox/Checkbox.test.tsx
@@ -3,7 +3,7 @@ import * as renderer from 'react-test-renderer';
 import { mount, ReactWrapper } from 'enzyme';
 
 import { Checkbox } from './Checkbox';
-import { IRefObject } from 'office-ui-fabric-react/lib/Utilities';
+import { IRefObject } from '../../Utilities';
 import { ICheckbox } from './Checkbox.types';
 
 let checkbox: ICheckbox | undefined;