+++ packages/example-app-base/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @uifabric/example-app-base
 
-This log was last generated on Fri, 04 Sep 2020 12:28:23 GMT and should not be manually modified.
+This log was last generated on Sun, 27 Sep 2020 04:20:52 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [7.15.18](https://github.com/microsoft/fluentui/tree/@uifabric/example-app-base_v7.15.18)
+
+Sun, 27 Sep 2020 04:20:52 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@uifabric/example-app-base_v7.15.4..@uifabric/example-app-base_v7.15.18)
+
+### Patches
+
+- Replace dep `@uifabric/fluent-theme` with `@fluentui/theme`. ([PR #15248](https://github.com/microsoft/fluentui/pull/15248) by xgao@microsoft.com)
 
 ## [7.15.4](https://github.com/microsoft/fluentui/tree/@uifabric/example-app-base_v7.15.4)